Changed Default site-wide project configuration to use new form (MAD-1438)
